Add read permission to directories
The '751' permission is not particularly restrictive, and results
in many unnecessary calls to sudo to enter application dirs just to
read stuff.

I would prefer to make this '2750' and add each admin user to the
apps group, but this is a wider change.

Adjusting to 755 is low risk, and not really any less secure than the
status quo.